+++ core/src/main/java/org/testcontainers/containers/wait/strategy/WaitStrategy.java
@@ -1,12 +1,10 @@
 package org.testcontainers.containers.wait.strategy;
 
-import org.testcontainers.ContainerState;
-
 import java.time.Duration;
 
 public interface WaitStrategy {
 
-    void waitUntilReady(ContainerState containerState);
+    void waitUntilReady(WaitStrategyTarget waitStrategyTarget);
 
     WaitStrategy withStartupTimeout(Duration startupTimeout);
 }